Improve documentation related to Stack hooks
Reorganises the existing 'Stack hooks' content to bring forward how to get the hook (using the recommended method).

In three relevant places, tells users that the commands to install GHCup and the Haskell Toolchain can safely be run again at any time.

Introduces that the commands to install GHCup and the Haskell Toolchain use a `sh` script (true on all platforms).

Makes the steps to get the Stack hook manually relevant to all operating systems (not just Unix-like ones).

Updates the text for the passage of time: e.g. Stack 2.9.1 is no longer 'new'.

Uses language about Stack that is closer to the language that Stack's online documentation uses about itself.

Also fixes a closing </p> tag (present as <p>).